feat(suite): update UdevRules modal
